Don't shrink the executor count to non-zero number
Fixes #14.

When using dynamic allocation, a smaller executor count than
what is currently running might be requested. Since we don't
want to kill existing allocations (they have shuffle services,
as well as executors that are potentially still running tasks),
we won't reduce the number of executor task groups unless we
are asked to reduce it to 0.